+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Version 1.26 [20-09-2024]
+Added CVars step to the installer and config panel  
+Added important WeakAura to handle friendly nameplates  
+___
 ### Version 1.25 [19-09-2024]
 Updated BigWigs profile  
 Updated ElvUI blacklist